+++ test/repatch_setup_test.exs
@@ -3,6 +3,8 @@ defmodule RepatchSetupTest do
 
   alias Repatch.Looper
 
+  @moduletag skip: true
+
   setup do
     Repatch.restore_all()